requirements.txt
# Snowflake dependencies
snowflake-connector-python[pandas]>=3.0.0
snowflake-ml-python>=1.5.0

# ML libraries
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.3.0
xgboost>=1.7.0
joblib>=1.3.0

# Visualization and interpretation
matplotlib>=3.6.0
seaborn>=0.11.0
shap>=0.41.0

# Environment and configuration
python-dotenv>=1.0.0

# General utilities
pyyaml>=6.0
requests>=2.28.0
urllib3>=1.26.0

# Development and testing (optional)
pytest>=7.0.0
pytest-cov>=4.0.0

# Optional: Jupyter for development
jupyter>=1.0.0
ipykernel>=6.0.0 